requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
distro==1.9.0
docopt==0.6.2
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
Js2Py==0.74
openai==1.76.2
packaging==25.0
pipwin==0.5.2
PyAudio==0.2.14
pydantic==2.11.4
pydantic_core==2.33.2
pyjsparser==2.7.1
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.1.0
pyttsx3==2.98
pywin32==310
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.2
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
